moved to update pallet

+++ client/src/pages/PalletPage.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 
 import { useEffect } from 'react';
 import { useState } from 'react';
@@ -12,6 +12,9 @@ import { toast } from 'react-toastify';
 
 
 import { removePallet } from '../redux/features/pallet/palletSlice';
+import PalletItem from '../components/Pallet/PalletItem';
+
+
 
 
 
@@ -23,39 +26,91 @@ import { removePallet } from '../redux/features/pallet/palletSlice';
 
 const PalletPage = () => {
 
-	const [pallet, setPallet] = useState(null)
+	const [pallet, setPallet] = useState("")
+	const [isEdit, setIsEdit] = useState(false)
 
 	const navigate = useNavigate()
 	const params = useParams()
 	const dispatch = useDispatch()
 
 
+	const fetchPallet = useCallback(async () => {
+		const { data } = await axios.get(`/pallets/${params.id}`)
+		setPallet(data)
+	}, [params.id])
+
+	useEffect(() => {
+		fetchPallet()
+	}, [fetchPallet])
+
+	const removeAttempt = () => {
+		window.confirm("Удалить эту паллету?") && dispatch(removePallet(params.id))
+			&& toast('Паллета был удалена')
+			&& navigate('/pallets')
+	}
+
 
 	const removePalletHandler = () => {
 		try {
-			dispatch(removePallet(params.id))
-			toast('Паллета был удалена')
-			navigate('/pallets')
+			removeAttempt()
+
 		} catch (error) {
 			console.log(error)
 		}
+	}
+
+	const handlerEdit = () => {
+		setIsEdit(true);
+	}
+
+	const handlerSave = () => {
+		setIsEdit(false);
 	}
 
 
 
 
 
+
+
 	return (
-		<div>
-			<button className='flex justify-center items-center bg-gray-600 text-xs text-white rounded-sm py-2 px-4'>
-				<Link className='flex' to={'/'}>
-					Назад
-				</Link>
-			</button>
+		<div className='mx-auto w-3/4  shadow-lg shadow-slate-400 rounded-b-md'>
+
+
+			<div className='text-3xl bg-teal-500 w-full my-4 flex justify-center' >{pallet.title}</div>
+
+			<PalletItem pallet={pallet} isEdit={isEdit} />
+
+			<div className='flex justify-center  w-full  my-3'>
 
 
 
+				{isEdit ?
 
+					<button
+						className='text-xl text-white p-2 rounded-lg  m-3   bg-blue-600'
+						onClick={handlerSave}
+
+					>Сохранить</button>
+					:
+
+					<button
+						className='text-xl text-white p-2 rounded-lg  m-3   bg-blue-600'
+						onClick={handlerEdit}
+
+					>Редактировать</button>}
+
+
+
+				<button className='text-xl text-black p-2 rounded-lg  m-3   bg-white' >Очистить</button>
+
+
+
+				<button
+					className='text-xl text-white p-2 rounded-lg  m-3   bg-red-600'
+					onClick={removePalletHandler}
+				>Удалить</button>
+			</div>
 
 
 		</div>
